+++ src/app/features/order/_components/order-list/order-list.component.ts
@@ -7,6 +7,7 @@ import {MatSort} from '@angular/material/sort';
 import {MatSnackBar} from '@angular/material/snack-bar';
 import {MatDialog} from '@angular/material/dialog';
 import {OrderFormComponent} from '../order-form/order-form.component';
+import {DeleteConfirmationComponent} from '../../../../shared/delete-confirmation/delete-confirmation.component';
 
 @Component({
   selector: 'app-order-list',
@@ -75,11 +76,18 @@ export class OrderListComponent implements OnInit {
     });
   }
 
-
   delete(id): void {
-    this.orderService.delete(id).subscribe(response => {
-      this.ngOnInit();
+    const dialogRef = this.dialog.open(DeleteConfirmationComponent, {
+      width: '1000px',
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.orderService.delete(id).subscribe(response => {
+          this.getOrderList();
+        });
+      }
     });
   }
+
 
 }
